Split keeper runtime static imports
