+++ dashboard/src/components/keeper-detail-page.ts
@@ -5,7 +5,7 @@ import { ChevronLeft, ChevronRight } from 'lucide-preact'
 import type { Keeper } from '../types'
 import { route } from '../router'
 import { keepers } from '../store'
-import { selectKeeper } from '../keeper-runtime'
+import { selectKeeper } from '../keeper-actions'
 import { KeeperConfigPanel, loadKeeperConfig } from './keeper-config-panel'
 import { findKeeper } from '../lib/keeper-utils'
 import { resolveKeeperForDetail } from '../lib/keeper-detail-resolution'
