+++ dashboard/src/components/keeper-detail-state.ts
@@ -1,7 +1,7 @@
 import { signal } from '@preact/signals'
 import { navigate, route } from '../router'
 import type { Keeper, TabId } from '../types'
-import { hydrateKeeperStatus, selectKeeper } from '../keeper-runtime'
+import { hydrateKeeperStatus, selectKeeper } from '../keeper-actions'
 import { activeKeeperName } from '../keeper-state'
 import { registerKeeperTurnRefresh } from '../sse-store'
 import { loadKeeperConfig, resetKeeperConfig } from './keeper-config-state'
