+++ dashboard/src/components/keeper-workspace/keeper-workspace-roster.ts
@@ -19,7 +19,7 @@ import { useEffect, useState } from 'preact/hooks'
 import type { VNode } from 'preact'
 import { keepers } from '../../store'
 import { navigate } from '../../router'
-import { selectKeeper } from '../../keeper-runtime'
+import { selectKeeper } from '../../keeper-actions'
 import { keeperMobilePane } from '../keeper-detail-state'
 import { formatRelativeSec } from '../../lib/format-time'
 import { keeperActivityDisplay } from '../../lib/keeper-runtime-display'
